[feat] domain 과 entity 분리
